Migrate TimeTupleToTime tests to munit
